to_bytes and from_bytes fails for transformer-based pipeline
